Updates oai-ue with MSIN values with 9 digits and zero padding
In DT the msinStart value is set to 0000000001.
The oai-ue helm chart considers that a binary, and change that to
1 (removes the left zeros).
In the file _ue_sim_conf.tpl , the line
MSIN={{ add $startMSIN $i | quote }};  makes MSIN equals to 1
(instead of 0000000001).
That creates an error in the init container and forbids its
initialization.
The solution is to add zeros as a padding in a printf string,
but another issue is that MSIN in Germany has 10 digits.
The solution also makes a padding to 9 digits MSIN.